mcp-engineering/project-001/app/requirements.txt
fastapi
httpx
uvicorn[standard]
pydantic
mcp
python-dotenv
fastmcp
openai